# openapi-schema.yml
openapi: 3.0.3
info:
  title: NoCode Backend API
  version: 1.0.0
  description: API pour la plateforme NoCode - Création d'applications sans code
paths:
...
  /api/v1/runtime/apps/:
    get:
      operationId: runtime_apps_list
      description: Retourne la liste des applications générées par l'utilisateur courant.
      summary: Lister les applications générées
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - runtime
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGeneratedAppList'
          description: ''
        '403':
          description: Accès non autorisé
    post:
      operationId: runtime_apps_create
      description: Crée une nouvelle application générée.
      summary: Créer une nouvelle application
      tags:
      - runtime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedApp'
          description: ''
        '400':
          description: Données invalides
        '403':
          description: Accès non autorisé
  /api/v1/runtime/apps/{id}/:
    get:
      operationId: runtime_apps_retrieve
      description: Récupère les détails d'une application générée.
      summary: Récupérer une application
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Application générée.
        required: true
      tags:
      - runtime
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedApp'
          description: ''
        '404':
          description: Application non trouvée
    put:
      operationId: runtime_apps_update
      description: Met à jour les informations d'une application générée.
      summary: Mettre à jour une application
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Application générée.
        required: true
      tags:
      - runtime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedApp'
          description: ''
        '400':
          description: Données invalides
        '403':
          description: Accès non autorisé
        '404':
          description: Application non trouvée
    patch:
      operationId: runtime_apps_partial_update
      description: API endpoint qui permet de gérer les applications générées.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Application générée.
        required: true
      tags:
      - runtime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedGeneratedAppRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedGeneratedAppRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedGeneratedAppRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedApp'
          description: ''
    delete:
      operationId: runtime_apps_destroy
      description: Supprime une application générée.
      summary: Supprimer une application
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Application générée.
        required: true
      tags:
      - runtime
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
        '403':
          description: Accès non autorisé
        '404':
          description: Application non trouvée
  /api/v1/runtime/apps/{id}/deploy/:
    post:
      operationId: runtime_apps_deploy_create
      description: Lance le déploiement d'une application générée sur l'environnement
        cible.
      summary: Déployer une application
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Application générée.
        required: true
      tags:
      - runtime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentActionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DeploymentActionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DeploymentActionRequest'
      security:
      - jwtAuth: []
      responses:
        '202':
          description: Déploiement démarré avec succès
        '400':
          description: L'application n'est pas dans un état déployable
        '403':
          description: Accès non autorisé
        '404':
          description: Application non trouvée
        '500':
          description: Erreur lors du démarrage du déploiement
...
  /api/v1/runtime/apps/{id}/status/:
    get:
      operationId: runtime_apps_status_retrieve
      description: Récupère les informations de statut d'une application générée,
        y compris son état de déploiement.
      summary: Obtenir le statut d'une application
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Application générée.
        required: true
      tags:
      - runtime
      security:
      - jwtAuth: []
      responses:
        '200':
          description: Statut de l'application récupéré avec succès
        '403':
          description: Accès non autorisé
        '404':
          description: Application non trouvée
        '500':
          description: Erreur lors de la récupération du statut
...
components:
  schemas:
...
    DeploymentActionRequest:
      type: object
      description: Sérialiseur pour les actions de déploiement.
      properties:
        target:
          enum:
          - local
          - staging
          - production
          type: string
          default: local
          description: |-
            Cible de déploiement (local, staging, production)

            * `local` - local
            * `staging` - staging
            * `production` - production
        force:
          type: boolean
          default: false
          description: Forcer le déploiement même si aucun changement détecté
...
    GeneratedApp:
      type: object
      description: Sérialiseur pour les applications générées.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          title: Nom de l'application
          maxLength: 255
        project:
          allOf:
          - $ref: '#/components/schemas/Project'
          readOnly: true
        version:
          type: string
          maxLength: 50
        status:
          enum:
          - draft
          - generated
          - deployment_pending
          - deployment_failed
          - deployed
          - error
          - archived
          type: string
          description: |-
            * `draft` - Brouillon
            * `generated` - Générée
            * `deployment_pending` - Déploiement en attente
            * `deployment_failed` - Échec du déploiement
            * `deployed` - Déployée
            * `error` - Erreur
            * `archived` - Archivée
          readOnly: true
          title: Statut
        status_display:
          type: string
          readOnly: true
        deployment_target:
          enum:
          - local
          - docker
          - staging
          - production
          - kubernetes
          type: string
          description: |-
            * `local` - Local (fichiers seulement)
            * `docker` - Docker (recommandé)
            * `staging` - Staging (Docker)
            * `production` - Production (Kubernetes)
            * `kubernetes` - Kubernetes
          title: Environnement de déploiement
        deployment_target_display:
          type: string
          readOnly: true
        api_base_url:
          type: string
          format: uri
          title: URL de base de l'API
          maxLength: 200
        admin_url:
          type: string
          format: uri
          title: URL d'administration
          maxLength: 200
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: Créé le
        updated_at:
          type: string
          format: date-time
          readOnly: true
          title: Mis à jour le
        last_deployed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          title: Dernier déploiement
        last_deployment:
          type: string
          readOnly: true
        config:
          type: object
          additionalProperties: {}
          title: Configuration technique
    GeneratedAppRequest:
      type: object
      description: Sérialiseur pour les applications générées.
      properties:
        name:
          type: string
          title: Nom de l'application
          maxLength: 255
        project_tracking_id:
          type: string
          format: uuid
          writeOnly: true
        version:
          type: string
          minLength: 1
          maxLength: 50
        deployment_target:
          enum:
          - local
          - docker
          - staging
          - production
          - kubernetes
          type: string
          description: |-
            * `local` - Local (fichiers seulement)
            * `docker` - Docker (recommandé)
            * `staging` - Staging (Docker)
            * `production` - Production (Kubernetes)
            * `kubernetes` - Kubernetes
          title: Environnement de déploiement
        api_base_url:
          type: string
          format: uri
          title: URL de base de l'API
          maxLength: 200
        admin_url:
          type: string
          format: uri
          title: URL d'administration
          maxLength: 200
        config:
          type: object
          additionalProperties: {}
          title: Configuration technique
      required:
      - project_tracking_id
...
    PaginatedGeneratedAppList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/GeneratedApp'
...
    PatchedGeneratedAppRequest:
      type: object
      description: Sérialiseur pour les applications générées.
      properties:
        name:
          type: string
          title: Nom de l'application
          maxLength: 255
        project_tracking_id:
          type: string
          format: uuid
          writeOnly: true
        version:
          type: string
          minLength: 1
          maxLength: 50
        deployment_target:
          enum:
          - local
          - docker
          - staging
          - production
          - kubernetes
          type: string
          description: |-
            * `local` - Local (fichiers seulement)
            * `docker` - Docker (recommandé)
            * `staging` - Staging (Docker)
            * `production` - Production (Kubernetes)
            * `kubernetes` - Kubernetes
          title: Environnement de déploiement
        api_base_url:
          type: string
          format: uri
          title: URL de base de l'API
          maxLength: 200
        admin_url:
          type: string
          format: uri
          title: URL d'administration
          maxLength: 200
        config:
          type: object
          additionalProperties: {}
          title: Configuration technique
...
    Project:
      type: object
      properties:
        tracking_id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 255
        organization_name:
          type: string
          readOnly: true
        schema_name:
          type: string
          readOnly: true
        created_by_email:
          type: string
          readOnly: true
        created_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - name
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: http://localhost:8000
  description: Development server
- url: https://api.nocode.com
  description: Production server